test: backend - user
urls test separated
